modal css + rename modal files & folders
