+++ docker-compose.yml
@@ -1,4 +1,4 @@
-version: "3.9"
+version: "3.1"
 services:
   mongo:
     container_name: mongo
