+++ frontend/src/components/modal/share-modal/share-modal.js
@@ -1,7 +1,9 @@
 import React, {useState} from 'react';
-import ShareList from '../share-users';
+import ShareList from '../share-list/share-list';
 import Modal from 'react-modal';
 import styles from './share-modal.module.css';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { faTimes } from '@fortawesome/free-solid-svg-icons'
 
 Modal.setAppElement('#root')
 
@@ -15,9 +17,15 @@ const ShareModal = ({modalIsOpen, toggleModal}) => {
           isOpen={modalIsOpen}
           onRequestClose={closeModal}
           className={styles.modal}
-        >
-            <ShareList />
-            <button onClick={closeModal}>close</button>
+          overlayClassName={styles.Overlay}
+        >   
+            <div className={styles.buttonContainer}>   
+                <span className={styles.message}>Select users to share!</span>
+                <button className={styles.button} onClick={closeModal}><FontAwesomeIcon icon={faTimes} /></button>
+            </div>
+            <div className={styles.shareContainer}>
+                <ShareList />
+            </div>  
         </Modal>
     )
 }
